add +-6 coins on blue/red spaces on last 5 turns

+++ mp1/bowser.go
@@ -92,7 +92,7 @@ const (
 	BBBP4Win
 )
 
-func GetBowserMinigameCoinLoss(turn uint) int {
+func GetBowserMinigameCoinLoss(turn uint8) int {
 	if turn <= 9 {
 		return 10
 	} else if turn <= 19 {
